初步完成图片墙重构

+++ Metic/Main Classes/PictureWallViewController.h
@@ -23,8 +23,8 @@
 @property(nonatomic,strong)NSNumber *sequence;
 @property (nonatomic,strong) NSMutableArray* lefPhotos;
 @property (nonatomic,strong) NSMutableArray* rigPhotos;
-@property float leftH;
-@property float rightH;
+@property double leftH;
+@property double rightH;
 @property(nonatomic,strong)NSMutableArray *photo_list;//部分
 @property(nonatomic,strong)NSMutableArray *photo_list_all;//总
 @property(nonatomic,strong)NSMutableArray *photoPath_list;
